+++ src/pytest_cov/engine.py
@@ -83,9 +83,15 @@ class CovController(object):
                 self.sep(stream, ' ', '%s' % node_desc)
 
         # Produce terminal report if wanted.
-        if 'term' in self.cov_report or 'term-missing' in self.cov_report:
-            show_missing = ('term-missing' in self.cov_report) or None
-            total = self.cov.report(show_missing=show_missing, ignore_errors=True, file=stream)
+        if any(x in self.cov_report for x in ['term', 'term-missing', 'term-skip-covered']):
+            options = {
+                'show_missing': ('term-missing' in self.cov_report) or None,
+                'ignore_errors': True,
+                'file': stream,
+            }
+            if hasattr(coverage, 'version_info') and coverage.version_info[0] >= 4:
+                options.update({'skip_covered': ('term-skip-covered' in self.cov_report) or None})
+            total = self.cov.report(**options)
 
         # Produce annotated source code report if wanted.
         if 'annotate' in self.cov_report: